Reworked api for viewing all STI history, reworked login and registra…

+++ backend/src/main/java/swp391/com/backend/jpa/repository/appointments/AppointmentsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import swp391.com.backend.jpa.pojo.appointments.Appointment;
 
 @Repository
-public interface AppointmentsRepository extends JpaRepository<Appointment, Integer> {
+public interface AppointmentsRepository extends JpaRepository<Appointment, Long> {
 }
